+++ src/query/storages/common/table_meta/src/meta/mod.rs
@@ -37,8 +37,11 @@ pub use statistics::*;
 // export legacy versioned table meta types locally,
 // currently, used by versioned readers only
 pub(crate) use testing::*;
+pub use utils::trim_v5_object_prefix;
+pub use utils::try_extract_uuid_str_from_path;
 pub use utils::uuid_from_date_time;
 pub use utils::TableMetaTimestamps;
+pub use utils::V5_OBJET_KEY_PREFIX;
 pub(crate) use utils::*;
 pub use versions::testify_version;
 pub use versions::SegmentInfoVersion;
